+++ flexiznam/main.py
@@ -18,10 +18,12 @@ def _format_project(project_id, prm):
     return project_id
 
 
-def _lookup_project(project_id, prm):
+def lookup_project(project_id, prm=None):
     """
     Look up project name by hexadecimal id
     """
+    if prm is None:
+        prm = PARAMETERS
     try:
         proj = next(proj for proj, id in prm["project_ids"].items() if id == project_id)
         return proj
@@ -115,25 +117,27 @@ def add_mouse(
             mcms_username = PARAMETERS["mcms_username"]
         if mcms_animal_name is None:
             mcms_animal_name = mouse_name
-        mcms_info = dict(
-            mcms.get_mouse_df(
-                mouse_name=mcms_animal_name,
-                username=mcms_username,
-                password=mcms_password,
-            )
+        mcms_info = mcms.get_mouse_info(
+            mouse_name=mcms_animal_name,
+            username=mcms_username,
+            password=mcms_password,
         )
-        # format properly results
-        for k, v in mcms_info.items():
-            if type(v) != str:
-                mcms_info[k] = float(v)
-            else:
-                mcms_info[k] = v.strip()
-
+        # flatten alleles and colony
+        alleles = mcms_info.pop("alleles")
+        for gene in alleles:
+            gene_name = gene["allele"]["shortAlleleSymbol"].replace(" ", "_")
+            mcms_info[gene_name] = gene["genotype"]["name"]
+        colony = mcms_info.pop("colony")
+        mcms_info["colony_prefix"] = colony["colonyPrefix"]
+        if not mcms_info:
+            raise IOError(f"Could not get info for mouse {mouse_name} from MCMS")
         # update mouse_info with mcms_info but prioritise mouse_info for conflicts
         mouse_info = dict(mcms_info, **mouse_info)
 
     # add the genealogy info, which is just [mouse_name]
     mouse_info["genealogy"] = [mouse_name]
+    project_name = lookup_project(flexilims_session.project_id, PARAMETERS)
+    mouse_info["path"] = str(Path(project_name) / mouse_name)
     resp = flexilims_session.post(
         datatype="mouse",
         name=mouse_name,
@@ -756,6 +760,7 @@ def get_entity(
     If multiple entities on the database match the query, raise a
     :py:class:`flexiznam.errors.NameNotUniqueError`, if nothing matches returns `None`.
 
+    For best performance, provide the `id` of the entity and/or the `datatype`.
     Args:
         datatype (str): type of Flexylims entity to fetch, e.g. 'mouse', 'session',
             'recording', or 'dataset'. If None, will iterate on all datatype until the
@@ -781,6 +786,10 @@ def get_entity(
     """
 
     if datatype is None:
+        if id is None:
+            warnings.warn(
+                "No datatype specified, trying everything. Will be slow", UserWarning
+            )
         # datatype is not specify, try everything
         args = [
             datatype,
@@ -964,7 +973,7 @@ def get_datasets(
     if flexilims_session is None:
         flexilims_session = get_flexilims_session(project_id)
     else:
-        project_id = _lookup_project(flexilims_session.project_id, PARAMETERS)
+        project_id = lookup_project(flexilims_session.project_id, PARAMETERS)
     recordings = get_entities(
         datatype="recording",
         origin_id=origin_id,
@@ -984,7 +993,7 @@ def get_datasets(
             flexilims_session=flexilims_session,
         )
         datapaths = []
-        for (dataset_path, is_raw) in zip(datasets["path"], datasets["is_raw"]):
+        for dataset_path, is_raw in zip(datasets["path"], datasets["is_raw"]):
             prefix = (
                 PARAMETERS["data_root"]["raw"]
                 if is_raw == "yes"