+++ flexiznam/mcms/__init__.py
@@ -1,1 +1,0 @@
-from flexiznam.mcms.main import get_mouse_df